Fix : No more use Ccommend, commend.js , now we can search by card name
