+++ routes/commend.js
@@ -2,6 +2,5 @@ const express = require('express');
 const controller = require('../controller/CCommend');
 const router = express.Router();
 
-router.get('/', controller.getCardByCategory);
 
 module.exports = router;
