+++ routes/search.js
@@ -2,6 +2,7 @@ const express = require('express');
 const controller = require('../controller/CSearch');
 const router = express.Router();
 
-router.get('/', controller.getCardDetails);
+router.get('/search', controller.getCardDetails);
+router.get('/category', controller.getCardByCategory);
 
 module.exports = router;
